+++ client/examples/get_funds.rs
@@ -3,11 +3,7 @@
 
 //! cargo run --example get_funds --release
 
-use iota_client::{
-    secret::{mnemonic::MnemonicSecretManager, SecretManager},
-    utils::request_funds_from_faucet,
-    Client, Result,
-};
+use iota_client::{secret::SecretManager, utils::request_funds_from_faucet, Client, Result};
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -22,9 +18,8 @@ async fn main() -> Result<()> {
         .with_node(&node_url)? // Insert the node here
         .finish()?;
 
-    let secret_manager = SecretManager::Mnemonic(MnemonicSecretManager::try_from_mnemonic(
-        &std::env::var("NON_SECURE_USE_OF_DEVELOPMENT_MNEMONIC_1").unwrap(),
-    )?);
+    let secret_manager =
+        SecretManager::try_from_mnemonic(&std::env::var("NON_SECURE_USE_OF_DEVELOPMENT_MNEMONIC_1").unwrap())?;
 
     let addresses = client
         .get_addresses(&secret_manager)