+++ client/examples/output/basic.rs
@@ -12,7 +12,7 @@ use iota_client::{
         },
         BasicOutputBuilder, Feature,
     },
-    secret::{mnemonic::MnemonicSecretManager, SecretManager},
+    secret::SecretManager,
     utils::request_funds_from_faucet,
     Client, Result,
 };
@@ -32,9 +32,8 @@ async fn main() -> Result<()> {
     // Create a client instance.
     let client = Client::builder().with_node(&node_url)?.finish()?;
 
-    let secret_manager = SecretManager::Mnemonic(MnemonicSecretManager::try_from_mnemonic(
-        &std::env::var("NON_SECURE_USE_OF_DEVELOPMENT_MNEMONIC_1").unwrap(),
-    )?);
+    let secret_manager =
+        SecretManager::try_from_mnemonic(&std::env::var("NON_SECURE_USE_OF_DEVELOPMENT_MNEMONIC_1").unwrap())?;
 
     let token_supply = client.get_token_supply().await?;