+++ client/tests/signing/basic.rs
@@ -16,7 +16,7 @@ use iota_client::{
         unlock::{SignatureUnlock, Unlock},
     },
     constants::{HD_WALLET_TYPE, SHIMMER_COIN_TYPE, SHIMMER_TESTNET_BECH32_HRP},
-    secret::{mnemonic::MnemonicSecretManager, SecretManage, SecretManageExt, SecretManager},
+    secret::{SecretManage, SecretManageExt, SecretManager},
     Client, Result,
 };
 
@@ -24,8 +24,7 @@ use crate::{build_inputs, build_outputs, Build::Basic};
 
 #[tokio::test]
 async fn single_ed25519_unlock() -> Result<()> {
-    let secret_manager =
-        SecretManager::Mnemonic(MnemonicSecretManager::try_from_mnemonic(&Client::generate_mnemonic()?)?);
+    let secret_manager = SecretManager::try_from_mnemonic(&Client::generate_mnemonic()?)?;
 
     let bech32_address_0 = &secret_manager
         .generate_addresses(SHIMMER_COIN_TYPE, 0, 0..1, false, None)
@@ -113,8 +112,7 @@ async fn single_ed25519_unlock() -> Result<()> {
 
 #[tokio::test]
 async fn ed25519_reference_unlocks() -> Result<()> {
-    let secret_manager =
-        SecretManager::Mnemonic(MnemonicSecretManager::try_from_mnemonic(&Client::generate_mnemonic()?)?);
+    let secret_manager = SecretManager::try_from_mnemonic(&Client::generate_mnemonic()?)?;
 
     let bech32_address_0 = &secret_manager
         .generate_addresses(SHIMMER_COIN_TYPE, 0, 0..1, false, None)
@@ -248,8 +246,7 @@ async fn ed25519_reference_unlocks() -> Result<()> {
 
 #[tokio::test]
 async fn two_signature_unlocks() -> Result<()> {
-    let secret_manager =
-        SecretManager::Mnemonic(MnemonicSecretManager::try_from_mnemonic(&Client::generate_mnemonic()?)?);
+    let secret_manager = SecretManager::try_from_mnemonic(&Client::generate_mnemonic()?)?;
 
     let bech32_address_0 = &secret_manager
         .generate_addresses(SHIMMER_COIN_TYPE, 0, 0..1, false, None)